Automount root device and ocfs2 support
mkinitrd will now create a ocfs2 data image instead of ext4. It will
also look for a cpio archive with the name of the host in it and add
a stanza for it in the extlinux.conf file.

srvcd will now look for a disk with a label same as the hostname and
mount it on /vol. It will then look for a boot sub-direcotry in /vol
and bind mount it on /boot.