Libarchive and ramdisk support
srvcd will create the cpio data archive directly instead of calling
the mkinitrd script to do it.

rdinit now uses the rxdsk and zram modules instead of the loop module.
It will setup the root filesystem on rxd0 instead of loop0 and
the data filesystem on zram0 instead of loop1.