LZMA support and cell data storage changes
When saving host data the device file is now first compressed
using LZMA2/XZ compression algorithm before adding it to the
cpio archive.

Cell data is now stored on disk in the /vol filesystem instead of
in memory.

The host data is still kept in memory but on a partitioned loop
device sitting on the zram device.